增加文艺片选项，先写test，后修改实现类

+++ src/main/java/rentalstore/Customer.java
@@ -55,9 +55,9 @@ public class Customer {
         return totalAmount;
     }
 
-    public int getTotalFrequentRenterPoints(){
+    public double getTotalFrequentRenterPoints(){
         Enumeration rentals = this.rentals.elements();
-        int frequentRenterPoints = 0;
+        double frequentRenterPoints = 0;
         while (rentals.hasMoreElements()) {
             Rental each = (Rental) rentals.nextElement();
             frequentRenterPoints = each.getFrequentRentalPoints();
